Desenvolvendo ClienteController

+++ BlueBank-API/src/main/java/com/blubank/repository/ClienteRepository.java
@@ -3,7 +3,7 @@ package com.blubank.repository;
 import com.blubank.model.Cliente;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ClienteRepository  extends JpaRepository<Cliente, Integer>{
+																   //Alteração para Long porque não estava puxando os dados como Integer.
+public interface ClienteRepository  extends JpaRepository<Cliente, Long>{
 
 }
